Time: 0 ms (100%), Space: 25.7 MB (85.94%) - LeetHub

+++ 2965-find-missing-and-repeated-values/2965-find-missing-and-repeated-values.cpp
@@ -3,10 +3,9 @@ public:
     vector<int> findMissingAndRepeatedValues(vector<vector<int>>& grid) {
         int n = grid.size();
         long long nSquare = n*n;
-
         long long actualSum = 0;
         long long actualSquareSum = 0;
-
+        
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < n; j++) {
                 actualSum += grid[i][j];
@@ -14,19 +13,18 @@ public:
             }
         }
 
-        
-        long long expectedSum = nSquare * (nSquare + 1) /2;
-        long long expectedSquareSum = nSquare * (nSquare + 1) * (2* nSquare + 1) / 6;
-
+        long long expectedSum = nSquare * (nSquare + 1) / 2;
+        long long expectedSquareSum = nSquare * (nSquare + 1) * (2 * nSquare + 1) / 6;
         //A-B
-        long long diffAB = actualSum - expectedSum;
+        int diffAB = actualSum - expectedSum;
         //A2-B2
-        long long diffSquareAB = actualSquareSum - expectedSquareSum;
+        int diffSquareAB = actualSquareSum - expectedSquareSum;
         //A+B
-        long long sumAB = diffSquareAB / diffAB;
+        int sumAB = diffSquareAB/diffAB;
 
         int a = (diffAB + sumAB) /2;
-        int b = (sumAB - diffAB) /2;
+        int b = (sumAB - diffAB) / 2;
+
         return {a,b};
     }
 };
